requirements.txt
flask-restx==0.2.0
Flask==1.1.2
Werkzeug==1.0.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.5
pytest==5.4.1
gunicorn==20.0.4
Flask-Cors==3.0.8  